[Console] Extract customizable BaseOutputStyle and reusable MessageBlock
