+++ src/Symfony/Component/Console/Output/OutputInterface.php
@@ -45,7 +45,7 @@ interface OutputInterface
      * @param string|iterable $messages The message as an iterable of strings or a single string
      * @param int             $options  A bitmask of options (one of the OUTPUT or VERBOSITY constants), 0 is considered the same as self::OUTPUT_NORMAL | self::VERBOSITY_NORMAL
      */
-    public function writeln($messages, int $options = 0);
+    public function writeln($messages, int $options = self::OUTPUT_NORMAL);
 
     /**
      * Sets the verbosity of the output.
